в шапку страницы добавлено распредление кнопок по ролям(user/organizer)
